requirements.txt
Flask==3.1.1
spacy==3.8.7
transformers==4.54.0
torch==2.7.1
pytesseract==0.3.13
opencv-python==4.12.0.88
rake-nltk==1.0.6
nltk==3.9.1
python-dotenv==1.1.1
requests==2.32.4
gunicorn==23.0.0
sentencepiece==0.2.0
protobuf==6.31.1
